refactor(insights-ui): remove unused synchronous stock report routes
These 8 per-report routes ran the LLM synchronously inside the request and were
only reachable from the old "Generate Synchronously" flow, which was removed when
generation became background-only. Nothing in the UI calls them anymore.

- Delete the 8 routes: business-and-moat, competition, fair-value, final-summary,
  financial-analysis, future-growth, management-team, past-performance.
- Delete llm-selection-utils.ts (only those routes used it).
- Remove the now-dead synchronous helpers from useGenerateReports
  (generateAnalysis / generateAllReportsForTicker / generateReportsSynchronously
  and the postAnalysis hook) and revert the AnalysisRequest llmProvider/model
  fields added for them.
- Relocate the FinalSummaryResponse / CompetitionAnalysisResponse types (used by
  the still-present save-json-report route) into analysis-factors-types.ts.

The `*-data` read routes and investor-analysis are untouched.
