--- insights-ui/src/app/api/[spaceId]/tickers-v1/exchange/[exchange]/[ticker]/business-and-moat/route.ts
@@ -1,63 +1,0 @@
-import { LLMFactorAnalysisResponse, TickerAnalysisResponse } from '@/types/public-equity/analysis-factors-types';
-import { TickerAnalysisCategory } from '@/types/ticker-typesv1';
-import { getLLMResponseForPromptViaInvocation } from '@/util/get-llm-response';
-import { parseLlmSelectionFromRequest } from '@/utils/analysis-reports/llm-selection-utils';
-import { fetchAnalysisFactors, fetchTickerRecordBySymbolAndExchangeWithIndustryAndSubIndustry } from '@/utils/analysis-reports/get-report-data-utils';
-import { saveBusinessAndMoatFactorAnalysisResponse } from '@/utils/analysis-reports/save-report-utils';
-import { prepareBusinessAndMoatInputJson } from '@/utils/analysis-reports/report-input-json-utils';
-import { ensureStockAnalyzerDataIsFresh, extractKpisDataForAnalysis } from '@/utils/stock-analyzer-scraper-utils';
-import { withAdminOrToken } from '@/app/api/helpers/withAdminOrToken';
-import { KoalaGainsJwtTokenPayload } from '@/types/auth';
-import { NextRequest } from 'next/server';
-
-async function postHandler(
-  req: NextRequest,
-  _userContext: KoalaGainsJwtTokenPayload | null,
-  { params }: { params: Promise<{ spaceId: string; ticker: string; exchange: string }> }
-): Promise<TickerAnalysisResponse> {
-  const { spaceId, ticker, exchange } = await params;
-
-  // Optional LLM provider/model chosen in the report-generation UI (falls back to defaults).
-  const { llmProvider, model } = await parseLlmSelectionFromRequest(req);
-
-  // Get ticker from DB
-  const tickerRecord = await fetchTickerRecordBySymbolAndExchangeWithIndustryAndSubIndustry(ticker.toUpperCase(), exchange.toUpperCase());
-
-  // Ensure stock analyzer data is fresh
-  const scraperInfo = await ensureStockAnalyzerDataIsFresh(tickerRecord);
-
-  // Get analysis factors for BusinessAndMoat category
-  const analysisFactors = await fetchAnalysisFactors(tickerRecord, TickerAnalysisCategory.BusinessAndMoat);
-
-  // Extract KPIs data for analysis
-  const kpisData = extractKpisDataForAnalysis(scraperInfo);
-
-  // Prepare input for the prompt
-  const inputJson = prepareBusinessAndMoatInputJson(tickerRecord, analysisFactors, kpisData);
-
-  // Call the LLM
-  const result = await getLLMResponseForPromptViaInvocation({
-    spaceId,
-    inputJson,
-    promptKey: 'US/public-equities-v1/business-moat',
-    requestFrom: 'ui',
-    llmProvider,
-    model,
-  });
-
-  if (!result) {
-    throw new Error('Failed to get response from LLM');
-  }
-
-  const response = result.response as LLMFactorAnalysisResponse;
-
-  // Save the analysis response using the utility function
-  await saveBusinessAndMoatFactorAnalysisResponse(ticker.toLowerCase(), tickerRecord.exchange, response, TickerAnalysisCategory.BusinessAndMoat);
-
-  return {
-    success: true,
-    invocationId: result.invocationId,
-  };
-}
-
-export const POST = withAdminOrToken<TickerAnalysisResponse>(postHandler);
