--- insights-ui/src/app/api/[spaceId]/tickers-v1/exchange/[exchange]/[ticker]/competition/route.ts
@@ -1,64 +1,0 @@
-import { getLLMResponseForPromptViaInvocation } from '@/util/get-llm-response';
-import { parseLlmSelectionFromRequest } from '@/utils/analysis-reports/llm-selection-utils';
-import { withAdminOrToken } from '@/app/api/helpers/withAdminOrToken';
-import { KoalaGainsJwtTokenPayload } from '@/types/auth';
-import { NextRequest } from 'next/server';
-import { TickerAnalysisResponse } from '@/types/public-equity/analysis-factors-types';
-import { saveCompetitionAnalysisResponse } from '@/utils/analysis-reports/save-report-utils';
-import { fetchTickerRecordBySymbolAndExchangeWithIndustryAndSubIndustry } from '@/utils/analysis-reports/get-report-data-utils';
-import { prepareBaseTickerInputJson } from '@/utils/analysis-reports/report-input-json-utils';
-
-export interface CompetitionAnalysisResponse {
-  summary: string;
-  overallAnalysisDetails: string;
-  competitionAnalysisArray: Array<{
-    companyName: string;
-    companySymbol?: string;
-    exchangeSymbol?: string;
-    exchangeName?: string;
-    detailedComparison: string;
-  }>;
-}
-
-async function postHandler(
-  req: NextRequest,
-  _userContext: KoalaGainsJwtTokenPayload | null,
-  { params }: { params: Promise<{ spaceId: string; ticker: string; exchange: string }> }
-): Promise<TickerAnalysisResponse> {
-  const { spaceId, ticker, exchange } = await params;
-
-  // Optional LLM provider/model chosen in the report-generation UI (falls back to defaults).
-  const { llmProvider, model } = await parseLlmSelectionFromRequest(req);
-
-  // Get ticker from DB
-  const tickerRecord = await fetchTickerRecordBySymbolAndExchangeWithIndustryAndSubIndustry(ticker.toUpperCase(), exchange.toUpperCase());
-
-  // Prepare input for the prompt (uses competition-input.schema.yaml)
-  const inputJson = prepareBaseTickerInputJson(tickerRecord);
-
-  // Call the LLM
-  const result = await getLLMResponseForPromptViaInvocation({
-    spaceId,
-    inputJson,
-    promptKey: 'US/public-equities-v1/competition',
-    requestFrom: 'ui',
-    llmProvider,
-    model,
-  });
-
-  if (!result) {
-    throw new Error('Failed to get response from LLM');
-  }
-
-  const response = result.response as CompetitionAnalysisResponse;
-
-  // Save the analysis response using the utility function
-  await saveCompetitionAnalysisResponse(ticker.toLowerCase(), tickerRecord.exchange, response);
-
-  return {
-    success: true,
-    invocationId: result.invocationId,
-  };
-}
-
-export const POST = withAdminOrToken<TickerAnalysisResponse>(postHandler);
