--- insights-ui/src/app/api/[spaceId]/tickers-v1/exchange/[exchange]/[ticker]/final-summary/route.ts
@@ -1,58 +1,0 @@
-import { getLLMResponseForPromptViaInvocation } from '@/util/get-llm-response';
-import { parseLlmSelectionFromRequest } from '@/utils/analysis-reports/llm-selection-utils';
-import { fetchTickerRecordBySymbolAndExchangeWithAnalysisData } from '@/utils/analysis-reports/get-report-data-utils';
-import { withAdminOrToken } from '@/app/api/helpers/withAdminOrToken';
-import { KoalaGainsJwtTokenPayload } from '@/types/auth';
-import { NextRequest } from 'next/server';
-import { TickerAnalysisResponse } from '@/types/public-equity/analysis-factors-types';
-import { saveFinalSummaryResponse } from '@/utils/analysis-reports/save-report-utils';
-import { prepareFinalSummaryInputJson } from '@/utils/analysis-reports/report-input-json-utils';
-
-export interface FinalSummaryResponse {
-  finalSummary: string;
-  metaDescription: string;
-  aboutReport: string;
-}
-
-async function postHandler(
-  req: NextRequest,
-  _userContext: KoalaGainsJwtTokenPayload | null,
-  { params }: { params: Promise<{ spaceId: string; ticker: string; exchange: string }> }
-): Promise<TickerAnalysisResponse> {
-  const { spaceId, ticker, exchange } = await params;
-
-  // Optional LLM provider/model chosen in the report-generation UI (falls back to defaults).
-  const { llmProvider, model } = await parseLlmSelectionFromRequest(req);
-
-  // Get ticker from DB with all related analysis data
-  const tickerRecord = await fetchTickerRecordBySymbolAndExchangeWithAnalysisData(ticker.toUpperCase(), exchange.toUpperCase());
-
-  // Prepare input for the prompt
-  const inputJson = prepareFinalSummaryInputJson(tickerRecord);
-
-  // Call the LLM
-  const result = await getLLMResponseForPromptViaInvocation({
-    spaceId,
-    inputJson,
-    promptKey: 'US/public-equities-v1/final-summary',
-    requestFrom: 'ui',
-    llmProvider,
-    model,
-  });
-
-  if (!result) {
-    throw new Error('Failed to get response from LLM');
-  }
-
-  const response = result.response as FinalSummaryResponse;
-
-  // Save the final summary response using the utility function
-  await saveFinalSummaryResponse(ticker.toLowerCase(), tickerRecord.exchange, response.finalSummary, response.metaDescription, response.aboutReport);
-
-  return {
-    success: true,
-    invocationId: result.invocationId,
-  };
-}
-
-export const POST = withAdminOrToken<TickerAnalysisResponse>(postHandler);
