--- insights-ui/src/app/api/[spaceId]/tickers-v1/exchange/[exchange]/[ticker]/financial-analysis/route.ts
@@ -1,63 +1,0 @@
-import { TickerAnalysisCategory } from '@/types/ticker-typesv1';
-import { getLLMResponseForPromptViaInvocation } from '@/util/get-llm-response';
-import { parseLlmSelectionFromRequest } from '@/utils/analysis-reports/llm-selection-utils';
-import { fetchAnalysisFactors, fetchTickerRecordBySymbolAndExchangeWithIndustryAndSubIndustry } from '@/utils/analysis-reports/get-report-data-utils';
-import { saveFinancialAnalysisFactorAnalysisResponse } from '@/utils/analysis-reports/save-report-utils';
-import { prepareFinancialAnalysisInputJson } from '@/utils/analysis-reports/report-input-json-utils';
-import { ensureStockAnalyzerDataIsFresh, extractFinancialDataForAnalysis } from '@/utils/stock-analyzer-scraper-utils';
-import { withAdminOrToken } from '@/app/api/helpers/withAdminOrToken';
-import { KoalaGainsJwtTokenPayload } from '@/types/auth';
-import { NextRequest } from 'next/server';
-import { LLMFactorAnalysisResponse, TickerAnalysisResponse } from '@/types/public-equity/analysis-factors-types';
-
-async function postHandler(
-  req: NextRequest,
-  _userContext: KoalaGainsJwtTokenPayload | null,
-  { params }: { params: Promise<{ spaceId: string; ticker: string; exchange: string }> }
-): Promise<TickerAnalysisResponse> {
-  const { spaceId, ticker, exchange } = await params;
-
-  // Optional LLM provider/model chosen in the report-generation UI (falls back to defaults).
-  const { llmProvider, model } = await parseLlmSelectionFromRequest(req);
-
-  // Get ticker from DB
-  const tickerRecord = await fetchTickerRecordBySymbolAndExchangeWithIndustryAndSubIndustry(ticker.toUpperCase(), exchange.toUpperCase());
-
-  // Ensure stock analyzer data is fresh
-  const scraperInfo = await ensureStockAnalyzerDataIsFresh(tickerRecord);
-
-  // Get analysis factors for FinancialStatementAnalysis category
-  const analysisFactors = await fetchAnalysisFactors(tickerRecord, TickerAnalysisCategory.FinancialStatementAnalysis);
-
-  // Extract comprehensive financial data for analysis
-  const financialData = extractFinancialDataForAnalysis(scraperInfo);
-
-  // Prepare input for the prompt
-  const inputJson = prepareFinancialAnalysisInputJson(tickerRecord, analysisFactors, financialData);
-
-  // Call the LLM
-  const result = await getLLMResponseForPromptViaInvocation({
-    spaceId,
-    inputJson,
-    promptKey: 'US/public-equities-v1/financial-statements',
-    requestFrom: 'ui',
-    llmProvider,
-    model,
-  });
-
-  if (!result) {
-    throw new Error('Failed to get response from LLM');
-  }
-
-  const response = result.response as LLMFactorAnalysisResponse;
-
-  // Save the analysis response using the utility function
-  await saveFinancialAnalysisFactorAnalysisResponse(ticker.toLowerCase(), tickerRecord.exchange, response, TickerAnalysisCategory.FinancialStatementAnalysis);
-
-  return {
-    success: true,
-    invocationId: result.invocationId,
-  };
-}
-
-export const POST = withAdminOrToken<TickerAnalysisResponse>(postHandler);
