+++ insights-ui/src/hooks/useGenerateReports.ts
@@ -1,7 +1,6 @@
 import { GenerationRequestPayload, TickerIdentifier } from '@/app/api/[spaceId]/tickers-v1/generation-requests/route';
 import { KoalaGainsSpaceId } from '@/types/koalaGainsConstants';
 import { LLMProvider } from '@/types/llmConstants';
-import { AnalysisRequest, TickerAnalysisResponse } from '@/types/public-equity/analysis-factors-types';
 import { InvestorKey, InvestorTypes, ReportType } from '@/types/ticker-typesv1';
 import { usePostData } from '@dodao/web-core/ui/hooks/fetch/usePostData';
 import getBaseUrl from '@dodao/web-core/utils/api/getBaseURL';
@@ -42,101 +41,12 @@ export const reportTypes: ReportTypeInfo[] = [
 export const useGenerateReports = () => {
   const [isGenerating, setIsGenerating] = useState<boolean>(false);
 
-  // Direct/synchronous analysis generation
-  const { postData: postAnalysis, loading: analysisLoading } = usePostData<TickerAnalysisResponse, AnalysisRequest>({
-    successMessage: 'Analysis generation started successfully!',
-    errorMessage: 'Failed to generate analysis.',
-  });
-
-  // Background batch generation
+  // Background batch generation. All stock report generation now goes through
+  // background generation requests (the synchronous per-report routes were removed).
   const { postData: postRequest, loading: requestLoading } = usePostData<unknown, GenerationRequestPayload[]>({
     successMessage: 'Background generation request created successfully!',
     errorMessage: 'Failed to create background generation request.',
   });
-
-  /** --- helpers (internal) --- */
-
-  const postAnalysisTo = async (url: string, payload: AnalysisRequest): Promise<TickerAnalysisResponse | undefined> => {
-    return postAnalysis(url, payload);
-  };
-
-  const generateAnalysis = async (
-    analysisType: ReportType,
-    ticker: TickerIdentifier,
-    onReportGenerated?: (ticker: TickerIdentifier) => void,
-    llmSelection?: ReportLlmSelection
-  ): Promise<void> => {
-    if (!ticker || !ticker.symbol || !ticker.exchange) return;
-    try {
-      const payload: AnalysisRequest = { llmProvider: llmSelection?.llmProvider, model: llmSelection?.model };
-      const result = await postAnalysisTo(
-        `${getBaseUrl()}/api/${KoalaGainsSpaceId}/tickers-v1/exchange/${ticker.exchange}/${ticker.symbol}/${analysisType}`,
-        payload
-      );
-      if (result && onReportGenerated) onReportGenerated(ticker);
-    } catch (err) {
-      console.error(`Error generating ${analysisType} for ${ticker.symbol}-${ticker.exchange}:`, err);
-    }
-  };
-
-  /** --- public API --- */
-
-  /**
-   * Generate all reports for a single ticker (synchronously, respecting dependencies).
-   */
-  const generateAllReportsForTicker = async (
-    ticker: TickerIdentifier,
-    onReportGenerated?: (ticker: TickerIdentifier) => void,
-    llmSelection?: ReportLlmSelection
-  ): Promise<void> => {
-    if (!ticker || !ticker.symbol || !ticker.exchange) return;
-
-    setIsGenerating(true);
-    try {
-      const sequence: ReportType[] = [
-        ReportType.FINANCIAL_ANALYSIS,
-        ReportType.COMPETITION, // must precede subsequent steps
-        ReportType.BUSINESS_AND_MOAT,
-        ReportType.FAIR_VALUE,
-        ReportType.PAST_PERFORMANCE,
-        ReportType.FUTURE_GROWTH,
-        ReportType.FINAL_SUMMARY,
-      ];
-
-      for (const step of sequence) {
-        await generateAnalysis(step, ticker, onReportGenerated, llmSelection);
-        await new Promise((r) => setTimeout(r, 1_000));
-      }
-    } finally {
-      setIsGenerating(false);
-    }
-  };
-
-  /**
-   * Generate selected report types synchronously for multiple tickers.
-   * Each ticker processes its selected steps in sequence (to respect deps).
-   */
-  const generateReportsSynchronously = async (
-    tickers: TickerIdentifier[],
-    selectedReportTypes: ReportType[],
-    onReportGenerated?: (ticker: TickerIdentifier) => void,
-    llmSelection?: ReportLlmSelection
-  ): Promise<void> => {
-    if (tickers.length === 0 || selectedReportTypes.length === 0) return;
-
-    setIsGenerating(true);
-    try {
-      const perTicker = tickers.map(async (ticker) => {
-        for (const reportType of selectedReportTypes) {
-          await generateAnalysis(reportType, ticker, onReportGenerated, llmSelection);
-          await new Promise((r) => setTimeout(r, 1_000));
-        }
-      });
-      await Promise.all(perTicker);
-    } finally {
-      setIsGenerating(false);
-    }
-  };
 
   /**
    * Batch background generation for specific report types and tickers.
@@ -297,16 +207,13 @@ export const useGenerateReports = () => {
 
   return {
     /** exports used by UI */
-    generateAllReportsForTicker,
-    generateReportsSynchronously,
     generateSpecificReportsInBackground,
     generateAllReportsInBackground,
     createFullBackgroundGenerationRequests,
     createFailedPartsOnlyGenerationRequests,
 
     /** state */
-    isGenerating: isGenerating || analysisLoading || requestLoading,
-    analysisLoading,
+    isGenerating: isGenerating || requestLoading,
     requestLoading,
   };
 };
